+++ lab_04/Button.h
@@ -18,21 +18,19 @@ public:
     Button(QWidget *parent = nullptr);
     ~Button() = default;
 
-    void set_floor(const ssize_t &floor);
+    void set_floor(const int &floor);
 
 public slots:
-
     void pressed();
     void unpressed();
 
 signals:
-
-    void floor_pressed(ssize_t floor);
+    void floor_pressed(int floor);
     void unpress_signal();
 
 private:
     button_status status;
-    ssize_t button_floor;
+    int button_floor;
 };
 
 #endif
